split and document formatters
